Adjust blog path format (#9)
Rename post and asset folder prefix to increasing numbers instead of
date. This change removes the need of renaming the files whilst still
maintaining a simple chronology at all times.

Load data that will be generated automatically, such as the date on build 
or by the CI from json and add generation script for publishing date.

Move static files to the correct folder to avoid missing files or
generated names.